move cmd/oohelperd lib code to internal/oohelperd
