+++ internal/cmd/oohelperd/main.go
@@ -7,27 +7,19 @@ import (
 	"fmt"
 	"net"
 	"net/http"
-	"net/http/cookiejar"
 	"net/http/pprof"
 	"os"
 	"os/signal"
 	"sync"
-	"sync/atomic"
 	"syscall"
 	"time"
 
 	"github.com/apex/log"
-	"github.com/ooni/probe-cli/v3/internal/model"
-	"github.com/ooni/probe-cli/v3/internal/netxlite"
+	"github.com/ooni/probe-cli/v3/internal/oohelperd"
 	"github.com/ooni/probe-cli/v3/internal/runtimex"
 	"github.com/ooni/probe-cli/v3/internal/version"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"golang.org/x/net/publicsuffix"
 )
-
-// maxAcceptableBodySize is the maximum acceptable body size for incoming
-// API requests as well as when we're measuring webpages.
-const maxAcceptableBodySize = 1 << 24
 
 var (
 	// apiEndpoint is the endpoint where we serve ooniprobe requests
@@ -58,16 +50,6 @@ var (
 	versionFlag = flag.Bool("version", false, "Prints version information on the stdout")
 )
 
-// newResolver creates a new [model.Resolver] suitable for serving
-// requests coming from ooniprobe clients.
-func newResolver(logger model.Logger) model.Resolver {
-	// Implementation note: pin to a specific resolver so we don't depend upon the
-	// default resolver configured by the box. Also, use an encrypted transport thus
-	// we're less vulnerable to any policy implemented by the box's provider.
-	resolver := netxlite.NewParallelDNSOverHTTPSResolver(logger, "https://dns.google/dns-query")
-	return resolver
-}
-
 // shutdown calls srv.Shutdown with a reasonably long timeout. The srv.Shutdown
 // function will immediately close any open listener and then will wait until
 // all pending connections are closed or the context has expired. By giving pending
@@ -79,90 +61,6 @@ func shutdown(srv *http.Server, wg *sync.WaitGroup) {
 	ctx, cancel := context.WithTimeout(context.Background(), 45*time.Second)
 	defer cancel()
 	srv.Shutdown(ctx)
-}
-
-// newCookieJar is the factory for constructing a new cookier jar.
-func newCookieJar() *cookiejar.Jar {
-	// Implementation note: the [cookiejar.New] function always returns a
-	// nil error; hence, it's safe here to use [runtimex.Try1].
-	return runtimex.Try1(cookiejar.New(&cookiejar.Options{
-		PublicSuffixList: publicsuffix.List,
-	}))
-}
-
-// newHTTPClientWithTransportFactory creates a new HTTP client.
-func newHTTPClientWithTransportFactory(
-	logger model.Logger,
-	txpFactory func(model.DebugLogger, model.Resolver) model.HTTPTransport,
-) model.HTTPClient {
-	// If the DoH resolver we're using insists that a given domain maps to
-	// bogons, make sure we're going to fail the HTTP measurement.
-	//
-	// The TCP measurements scheduler in ipinfo.go will also refuse to
-	// schedule TCP measurements for bogons.
-	//
-	// While this seems theoretical, as of 2022-08-28, I see:
-	//
-	//     % host polito.it
-	//     polito.it has address 192.168.59.6
-	//     polito.it has address 192.168.40.1
-	//     polito.it mail is handled by 10 mx.polito.it.
-	//
-	// So, it's better to consider this as a possible corner case.
-	reso := netxlite.MaybeWrapWithBogonResolver(
-		true, // enabled
-		newResolver(logger),
-	)
-
-	// fix: We MUST set a cookie jar for measuring HTTP. See
-	// https://github.com/ooni/probe/issues/2488 for additional
-	// context and pointers to the relevant measurements.
-	client := &http.Client{
-		Transport:     txpFactory(logger, reso),
-		CheckRedirect: nil,
-		Jar:           newCookieJar(),
-		Timeout:       0,
-	}
-
-	return netxlite.WrapHTTPClient(client)
-}
-
-// newHandler constructs the [handler] used by [main].
-func newHandler() *handler {
-	return &handler{
-		BaseLogger:        log.Log,
-		Indexer:           &atomic.Int64{},
-		MaxAcceptableBody: maxAcceptableBodySize,
-		Measure:           measure,
-
-		NewHTTPClient: func(logger model.Logger) model.HTTPClient {
-			return newHTTPClientWithTransportFactory(
-				logger,
-				netxlite.NewHTTPTransportWithResolver,
-			)
-		},
-
-		NewHTTP3Client: func(logger model.Logger) model.HTTPClient {
-			return newHTTPClientWithTransportFactory(
-				logger,
-				netxlite.NewHTTP3TransportWithResolver,
-			)
-		},
-
-		NewDialer: func(logger model.Logger) model.Dialer {
-			return netxlite.NewDialerWithoutResolver(logger)
-		},
-		NewQUICDialer: func(logger model.Logger) model.QUICDialer {
-			return netxlite.NewQUICDialerWithoutResolver(
-				netxlite.NewQUICListener(),
-				logger,
-			)
-		},
-		NewResolver: newResolver,
-		NewTLSHandshaker: func(logger model.Logger) model.TLSHandshaker {
-			return netxlite.NewTLSHandshakerStdlib(logger)
-		},
-	}
 }
 
 func main() {
@@ -194,7 +92,7 @@ func main() {
 	mux := http.NewServeMux()
 
 	// add the main oohelperd handler to the mux
-	mux.Handle("/", newHandler())
+	mux.Handle("/", oohelperd.NewHandler())
 
 	// create a listening server for serving ooniprobe requests
 	srv := &http.Server{Addr: *apiEndpoint, Handler: mux}
