+++ internal/oohelperd/dns.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 //
 // DNS measurements
