+++ internal/oohelperd/http.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 //
 // HTTP measurements
