+++ internal/oohelperd/http_test.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 import (
 	"context"
