+++ internal/oohelperd/logging.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 //
 // Logging code
