+++ internal/oohelperd/measure.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 //
 // Top-level measurement algorithm
@@ -24,7 +24,7 @@ type (
 
 // measure performs the measurement described by the request and
 // returns the corresponding response or an error.
-func measure(ctx context.Context, config *handler, creq *ctrlRequest) (*ctrlResponse, error) {
+func measure(ctx context.Context, config *Handler, creq *ctrlRequest) (*ctrlResponse, error) {
 	// create indexed logger
 	logger := &prefixLogger{
 		indexstr: fmt.Sprintf("<#%d> ", config.Indexer.Add(1)),
