+++ internal/oohelperd/metrics.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 //
 // Metrics definitions
