+++ internal/oohelperd/quic.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 //
 // QUIC handshake measurements
