+++ internal/oohelperd/tcptls_test.go
@@ -1,4 +1,4 @@
-package main
+package oohelperd
 
 import (
 	"testing"
